+++ libs/cardano-ledger-binary/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.9.0.0
 
-* Add `mapSparseField`, `mapSparseFieldA`, `mapSparseFieldOptional`, `mapSparseFieldGuarded`
+* Add `decodeAccA`, `failOnNull`, `failOnMempty`
 * Add `decodeSparseKeyed`
 * Add `decodeIntegralRational`
 * Add `decodeNonEmptySetLikeEnforceNoDuplicates`